+++ Engine/plugins/nsis.py
@@ -28,15 +28,15 @@ except ImportError:
 BYTE = c_ubyte
 WORD = c_ushort
 LONG = c_long
-DWORD = c_ulong
+DWORD = c_uint
 FLOAT = c_float
 LPBYTE = POINTER(c_ubyte)
 LPTSTR = POINTER(c_char)
 HANDLE = c_void_p
 PVOID = c_void_p
 LPVOID = c_void_p
-UINT_PTR = c_ulong
-SIZE_T = c_ulong
+UINT_PTR = c_uint
+SIZE_T = c_uint
 
 class StructNsisHeader(Structure):
     _pack_ = 1
